AG-8290 Move labelUtil.ts
