+++ packages/ag-charts-community/src/chart/labelUtil.ts
@@ -1,10 +1,10 @@
 import type { AnyFn, IsAny, RequireOptional } from 'ag-charts-core';
 import type { AgChartLabelStyleOptions, AgChartLabelStylerParams } from 'ag-charts-types';
 
-import type { Point } from '../../../scene/point';
-import type { Text } from '../../../scene/shape/text';
-import { mergeDefaults } from '../../../util/object';
-import type { Label } from '../../label';
+import type { Point } from '../scene/point';
+import type { Text } from '../scene/shape/text';
+import { mergeDefaults } from '../util/object';
+import type { Label } from './label';
 
 interface SeriesLike {
     id: string;
