+++ packages/ag-charts-community/src/chart/series/cartesian/barSeries.ts
@@ -31,6 +31,7 @@ import {
     processedDataIsAnimatable,
     valueProperty,
 } from '../../data/processors';
+import { adjustLabelPlacement, updateLabelNode } from '../../labelUtil';
 import type { CategoryLegendDatum, ChartLegendType } from '../../legend/legendDatum';
 import type { LegendSymbolOptions } from '../../legend/legendSymbol';
 import { type TooltipContent } from '../../tooltip/tooltip';
@@ -63,7 +64,6 @@ import {
     DEFAULT_CARTESIAN_DIRECTION_NAMES,
 } from './cartesianSeries';
 import { calculateDataDiff } from './diffUtil';
-import { adjustLabelPlacement, updateLabelNode } from './labelUtil';
 import { areScalingEqual } from './scaling';
 
 interface BarNodeLabelDatum extends Readonly<Point> {
