+++ packages/ag-charts-community/src/chart/series/cartesian/histogramSeries.ts
@@ -26,6 +26,7 @@ import {
     rowCountProperty,
     valueProperty,
 } from '../../data/processors';
+import { getLabelStyles } from '../../labelUtil';
 import type { CategoryLegendDatum, ChartLegendType } from '../../legend/legendDatum';
 import type { LegendSymbolOptions } from '../../legend/legendSymbol';
 import { type TooltipContent, type TooltipContentDataRow } from '../../tooltip/tooltip';
@@ -46,7 +47,6 @@ import {
 } from './cartesianSeries';
 import { calculateDataDiff } from './diffUtil';
 import { type HistogramNodeDatum, HistogramSeriesProperties } from './histogramSeriesProperties';
-import { getLabelStyles } from './labelUtil';
 import { addHitTestersToQuadtree, findQuadtreeMatch } from './quadtreeUtil';
 
 const defaultBinCount = 10;
