+++ packages/ag-charts-community/src/module-support.ts
@@ -68,6 +68,7 @@ export * from './chart/data/dataService';
 export * from './chart/data/processors';
 export * from './chart/data/aggregateFunctions';
 export * from './chart/updateService';
+export * from './chart/labelUtil';
 export * from './chart/layout/layoutManager';
 export * from './chart/interaction/animationManager';
 export * from './chart/interaction/contextMenuRegistry';
@@ -99,7 +100,6 @@ export * from './chart/series/cartesian/barUtil';
 export * from './chart/series/cartesian/areaUtil';
 export * from './chart/series/cartesian/diffUtil';
 export * from './chart/series/cartesian/markerUtil';
-export * from './chart/series/cartesian/labelUtil';
 export * from './chart/series/cartesian/pathUtil';
 export * from './chart/series/cartesian/quadtreeUtil';
 export * from './chart/series/cartesian/interpolationProperties';
